# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blasé Plaza Archives API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type SurveyDemolitionHorizon =
  (typeof SurveyDemolitionHorizon)[keyof typeof SurveyDemolitionHorizon];

export const SurveyDemolitionHorizon = {
  IMMINENT: "IMMINENT",
  "NEAR-TERM": "NEAR-TERM",
  PROJECTED: "PROJECTED",
  EXPIRED: "EXPIRED",
} as const;

export interface Survey {
  id: number;
  siteId: string;
  plazaName: string;
  location: string;
  surveyDate: string;
  demolitionHorizon: SurveyDemolitionHorizon;
  plazaType: string;
  architecturalStyle: string;
  parkingEntropy: string;
  shadeCoverage: string;
  signageDensity: string;
  vacancyRatio: string;
  pedestrianActivity: string;
  reportText: string;
  permitNo: string;
  permitType: string;
  permitIssueDate: string;
  documentRef: string;
  latitude?: number | null;
  longitude?: number | null;
  status: string;
  createdAt: string;
  lastSyncedAt?: string | null;
}

export type SurveyStatsByHorizon = {
  IMMINENT: number;
  "NEAR-TERM": number;
  PROJECTED: number;
  EXPIRED: number;
};

export interface SurveyStats {
  total: number;
  byHorizon: SurveyStatsByHorizon;
  lastUpdated: string | null;
}

export interface SyncResult {
  success: boolean;
  message: string;
  processed: number;
  errors: number;
}

export interface ErrorResponse {
  error: string;
}

export type GetSurveysParams = {
  horizon?: GetSurveysHorizon;
  search?: string;
  sortBy?: string;
  sortDir?: GetSurveysSortDir;
};

export type GetSurveysHorizon =
  (typeof GetSurveysHorizon)[keyof typeof GetSurveysHorizon];

export const GetSurveysHorizon = {
  IMMINENT: "IMMINENT",
  "NEAR-TERM": "NEAR-TERM",
  PROJECTED: "PROJECTED",
  EXPIRED: "EXPIRED",
} as const;

export type GetSurveysSortDir =
  (typeof GetSurveysSortDir)[keyof typeof GetSurveysSortDir];

export const GetSurveysSortDir = {
  asc: "asc",
  desc: "desc",
} as const;