# lib/api-zod/src/generated/types/survey.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blasé Plaza Archives API
 * OpenAPI spec version: 0.1.0
 */
import type { SurveyDemolitionHorizon } from "./surveyDemolitionHorizon";

export interface Survey {
  id: number;
  siteId: string;
  plazaName: string;
  location: string;
  surveyDate: string;
  demolitionHorizon: SurveyDemolitionHorizon;
  plazaType: string;
  architecturalStyle: string;
  parkingEntropy: string;
  shadeCoverage: string;
  signageDensity: string;
  vacancyRatio: string;
  pedestrianActivity: string;
  reportText: string;
  permitNo: string;
  permitType: string;
  permitIssueDate: string;
  documentRef: string;
  latitude?: number | null;
  longitude?: number | null;
  status: string;
  createdAt: string;
  lastSyncedAt?: string | null;
}