# lib/api-zod/src/generated/types/surveyStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blasé Plaza Archives API
 * OpenAPI spec version: 0.1.0
 */
import type { SurveyStatsByHorizon } from "./surveyStatsByHorizon";

export interface SurveyStats {
  total: number;
  byHorizon: SurveyStatsByHorizon;
  lastUpdated: string | null;
}