# lib/api-zod/src/generated/types/syncResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Blasé Plaza Archives API
 * OpenAPI spec version: 0.1.0
 */

export interface SyncResult {
  success: boolean;
  message: string;
  processed: number;
  errors: number;
}